# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TMCpos Textile POS/ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: lowStock
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}/rolls:
    get:
      operationId: getProductRolls
      tags: [products]
      summary: Get rolls for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of rolls
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductRoll"
    post:
      operationId: createProductRoll
      tags: [products]
      summary: Manually create a roll for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRollInput"
      responses:
        "201":
          description: Created roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductRoll"
  /products/{id}/rolls/{rollId}:
    patch:
      operationId: updateProductRoll
      tags: [products]
      summary: Update a product roll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: rollId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRollUpdateInput"
      responses:
        "200":
          description: Updated roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductRoll"
    delete:
      operationId: deleteProductRoll
      tags: [products]
      summary: Delete a product roll
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: rollId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers (Pelanggan)
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, categoryId, barcode, primaryUnit, secondaryUnit, lotNumber, rackLocation, costPricePerMeter, costPricePerRoll, pricePerMeter, pricePerRoll, rollStock, meterStock, minStock, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        barcode:
          type: string
        primaryUnit:
          type: string
        secondaryUnit:
          type: string
        lotNumber:
          type: string
        rackLocation:
          type: string
        pricePerMeter:
          type: number
        pricePerRoll:
          type: ["number", "null"]
        rollStock:
          type: number
        meterStock:
          type: number
        minStock:
          type: number
        isLowStock:
          type: boolean
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, categoryId]
      properties:
        name:
          type: string
        categoryId:
          type: integer
        barcode:
          type: string
        primaryUnit:
          type: string
        secondaryUnit:
          type: string
        lotNumber:
          type: string
        rackLocation:
          type: string
        costPricePerMeter:
          type: number
        costPricePerRoll:
          type: number
        pricePerMeter:
          type: number
        pricePerRoll:
          type: number
        rollStock:
          type: number
        meterStock:
          type: number
        minStock:
          type: number
    ProductRollInput:
      type: object
      required: [originalLength, currentLength]
      properties:
        barcode:
          type: string
        originalLength:
          type: number
        currentLength:
          type: number
    ProductRollUpdateInput:
      type: object
      properties:
        barcode:
          type: string
        originalLength:
          type: number
        currentLength:
          type: number
    ProductRoll:
      type: object
      required: [id, productId, barcode, originalLength, currentLength, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        barcode:
          type: string
        originalLength:
          type: number
        currentLength:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
        barcode:
          type: string
        primaryUnit:
          type: string
        secondaryUnit:
          type: string
        lotNumber:
          type: string
        rackLocation:
          type: string
        costPricePerMeter:
          type: number
        costPricePerRoll:
          type: number
        pricePerMeter:
          type: number
        pricePerRoll:
          type: number
        minStock:
          type: number
        rollStock:
          type: number
        meterStock:
          type: number
...